+++ app/Http/Controllers/TaskController.php
@@ -2,20 +2,24 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Course;
 use App\Models\Task;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 
 class TaskController extends Controller
 {
-    public function store(Request $request, $course)
+    use AuthorizesRequests;
+
+    public function store(Request $request, Course $course)
     {
         $taskData = $request->validate([
             'title' => ["required", "string", "max:255", "unique:tasks,title"],
             'content' => ["required", "string", "max:255"],
         ]);
 
-        $taskData["course_id"] = $course;
+        $taskData["course_id"] = $course->id;
 
         Task::query()->create($taskData);
 
@@ -26,6 +30,8 @@ class TaskController extends Controller
 
     public function update(Request $request, Task $task)
     {
+        $this->authorize('checkOwnership', $task);
+
         $taskData = $request->validate([
             'title' => ["nullable", "string", "max:255", Rule::unique('tasks')->ignore($task->id)],
             'content' => ["nullable", "string", "max:255"],
@@ -40,6 +46,8 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        $this->authorize('checkOwnership', $task);
+
         $task->delete();
 
         return response()->json([
